requirements.txt
Django==5.2
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
drf-spectacular==0.28.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
django-filter==25.1
django-debug-toolbar==5.1.0
python-dotenv==1.1.0
tensorflow==2.19.0
scikit-learn==1.6.1
pillow==11.2.1